Finish at home

+++ myoro_fitness/lib/helpers/conversion_helper.dart
@@ -1,3 +1,3 @@
 class ConversionHelper {
-  static int killojoulesToCalories(int kj) => kj ~/ 4.184;
+  static int killojoulesToCalories(double kj) => kj ~/ 4.184;
 }
