+++ myoro_fitness/lib/models/nutrient_model.dart
@@ -3,7 +3,7 @@ import 'package:myoro_fitness/enums/unit_enum.dart';
 
 class NutrientModel {
   final String name;
-  final int value;
+  final double value;
   final UnitEnum? unit;
 
   NutrientModel({
@@ -12,17 +12,17 @@ class NutrientModel {
     this.unit,
   });
 
-  set value(int value) => this.value = value;
+  set value(double value) => this.value = value;
   set unit(UnitEnum? unit) => this.unit = unit;
 
   NutrientModel.fromJSON(Map<String, dynamic> json)
       : name = NutrientEnum.values.firstWhere((element) => element.apiNutrientName == json['nutrientName']).title,
-        value = int.parse(json['nutrientNumber']),
+        value = double.parse(json['nutrientNumber']),
         unit = json['unitName'] != null ? UnitEnum.values.firstWhere((element) => element.notation == json['unitName']) : null;
 
   NutrientModel copyWith({
     String? name,
-    int? value,
+    double? value,
     UnitEnum? unit,
   }) =>
       NutrientModel(
